+++ yamlgraph/node_factory/llm_nodes.py
@@ -80,6 +80,7 @@ def create_node_function(
     # Get config values (node > defaults)
     temperature = node_config.get("temperature", defaults.get("temperature", 0.7))
     provider = node_config.get("provider", defaults.get("provider"))
+    model = node_config.get("model", defaults.get("model"))
     state_key = node_config.get("state_key", node_name)
     variable_templates = node_config.get("variables", {})
     requires = node_config.get("requires", [])
@@ -136,6 +137,7 @@ def create_node_function(
                     output_model=output_model,
                     temperature=temperature,
                     provider=use_provider,
+                    model=model,
                     graph_path=graph_path,
                     prompts_dir=prompts_dir,
                     prompts_relative=prompts_relative,